refactor(112070): :recycle: separa logica do componente
